Image example/requirements.txt
# Core machine learning and data science libraries
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.1.0
dill>=0.3.0

# Image processing
Pillow>=9.0.0

# # Data visualization and plotting
matplotlib>=3.5.0

# # Jupyter notebook support (if using the explore notebook)
# jupyter>=1.0.0
# ipykernel>=6.0.0

# # Optional: Additional useful libraries
# pandas>=1.4.0
# seaborn>=0.11.0